refactor: group toolbar components
