+++ packages/rich-text/src/Toolbar/components/EmbeddedEntityDropdownButton.tsx
@@ -3,8 +3,7 @@ import { Button, Menu } from '@contentful/f36-components';
 
 import { PlusIcon, ChevronDownIcon } from '@contentful/f36-icons';
 
-// TODO: this isn't a plugin. Move somewhere else
-interface EmbeddedEntityDropdownButtonProps {
+export interface EmbeddedEntityDropdownButtonProps {
   children: React.ReactNode;
   isDisabled: boolean | undefined;
   isOpen: boolean;
